requirements.txt
chromadb
python-dotenv